requirements.txt
# Ваши текущие зависимости
Pillow==9.0.0
openpyxl==3.0.10
tabulate==0.9.0
PyMuPDF>=1.23.0  # Для работы с PDF
schedule>=1.1.0
